requirements.txt
python-telegram-bot==20.3
nudenet==0.2.0
Pillow==10.0.0